refactor: extract data app membership page to multiple components
