+++ enterprise/frontend/src/metabase-enterprise/data_apps/admin/components/DataAppUserList/DataAppUserList.module.css
@@ -1,0 +1,3 @@
+.userTable tbody tr:not(:last-child) {
+  border-bottom: 1px solid var(--mb-color-border);
+}
